Add virtual TLS header
- Add virtual `tls.h` header for commonly used TLS headers
- Make `Record_Type` internal

#3324
