+++ src/examples/tls_proxy.cpp
@@ -2,11 +2,7 @@
 #include <botan/certstor.h>
 #include <botan/pk_keys.h>
 #include <botan/pkcs8.h>
-#include <botan/tls_callbacks.h>
-#include <botan/tls_client.h>
-#include <botan/tls_policy.h>
-#include <botan/tls_server.h>
-#include <botan/tls_session_manager_memory.h>
+#include <botan/tls.h>
 
 #include <memory>
