+++ src/lib/tls/tls.h
@@ -1,0 +1,23 @@
+/**
+ * Virtual header for commonly used TLS headers.
+ * (C) 2023 Jack Lloyd
+ *     2023 Philippe Lieser - Rohde & Schwarz Cybersecurity
+ *
+ * Botan is released under the Simplified BSD License (see license.txt)
+ */
+#ifndef BOTAN_TLS_H_
+#define BOTAN_TLS_H_
+
+#include <botan/credentials_manager.h>
+#include <botan/tls_callbacks.h>
+#include <botan/tls_client.h>
+#include <botan/tls_exceptn.h>
+#include <botan/tls_policy.h>
+#include <botan/tls_server.h>
+#include <botan/tls_session_manager.h>
+#include <botan/tls_session_manager_noop.h>
+#include <botan/tls_session_manager_memory.h>
+#include <botan/tls_session_manager_stateless.h>
+#include <botan/tls_session_manager_hybrid.h>
+
+#endif
