+++ src/lib/tls/tls12/tls_handshake_io.h
@@ -10,6 +10,7 @@
 
 #include <botan/tls_magic.h>
 #include <botan/tls_version.h>
+#include <botan/internal/tls_channel_impl.h>
 #include <functional>
 #include <vector>
 #include <deque>
