+++ src/lib/tls/tls13/tls_record_layer_13.h
@@ -16,6 +16,7 @@
 
 #include <botan/secmem.h>
 #include <botan/tls_magic.h>
+#include <botan/internal/tls_channel_impl.h>
 
 namespace Botan::TLS {
