+++ src/lib/tls/tls_magic.h
@@ -50,17 +50,6 @@ enum class Connection_Side {
    SERVER BOTAN_DEPRECATED("Use Connection_Side::Server") = Server,
 };
 
-enum class Record_Type: uint8_t {
-   Invalid            = 0,  // RFC 8446 (TLS 1.3)
-
-   ChangeCipherSpec   = 20,
-   Alert              = 21,
-   Handshake          = 22,
-   ApplicationData    = 23,
-
-   Heartbeat          = 24, // RFC 6520 (TLS 1.3)
-};
-
 enum class Handshake_Type {
    HelloRequest         = 0,
    ClientHello          = 1,
